# api-docs/openapi.json
{"openapi":"3.0.1","info":{"title":"School Management App API","description":"REST API for login, students, attendance, timetable, homework, exam results, notices, and fees","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/v1/users":{"get":{"tags":["Users"],"summary":"List users (paginated), optionally filtered by role","operationId":"list","parameters":[{"name":"role","in":"query","required":false,"schema":{"type":"string","enum":["ADMIN","TEACHER","PARENT"]}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageUserDto"}}}}}},"post":{"tags":["Users"],"summary":"Create a Teacher or Parent user account","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/v1/timetable":{"post":{"tags":["Timetable"],"summary":"Create a timetable entry","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TimetableCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TimetableDto"}}}}}}},"/api/v1/students":{"get":{"tags":["Students"],"summary":"List students (paginated)","operationId":"list_1","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageStudentDto"}}}}}},"post":{"tags":["Students"],"summary":"Create a student","operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StudentDto"}}}}}}},"/api/v1/notices":{"get":{"tags":["Notices"],"summary":"List notices visible to the current user, optionally filtered by target role","operationId":"list_2","parameters":[{"name":"role","in":"query","required":false,"schema":{"type":"string","enum":["ADMIN","TEACHER","PARENT","ALL"]}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageNoticeDto"}}}}}},"post":{"tags":["Notices"],"summary":"Post a notice","operationId":"create_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoticeCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NoticeDto"}}}}}}},"/api/v1/homework":{"post":{"tags":["Homework"],"summary":"Post homework for a class/section","operationId":"create_4","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HomeworkCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HomeworkDto"}}}}}}},"/api/v1/exam-results":{"post":{"tags":["Exam Results"],"summary":"Record an exam result (grade is computed server-side)","operationId":"create_5","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExamResultCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExamResultDto"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["Auth"],"summary":"Exchange a valid refresh token for a new token pair","operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Auth"],"summary":"Authenticate with email/password and receive access + refresh tokens","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/v1/attendance":{"post":{"tags":["Attendance"],"summary":"Mark (or update) attendance for a student on a given date","operationId":"mark","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttendanceMarkRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AttendanceDto"}}}}}}},"/api/v1/students/{id}":{"get":{"tags":["Students"],"summary":"Get a student by id (parents may only view their own child)","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StudentDto"}}}}}},"patch":{"tags":["Students"],"summary":"Update a student","operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StudentDto"}}}}}}},"/api/v1/fees/{id}":{"patch":{"tags":["Fees"],"summary":"Update a fee record's paid amount/status","operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeeUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FeeDto"}}}}}}},"/api/v1/timetable/{studentClass}/{section}":{"get":{"tags":["Timetable"],"summary":"View the weekly timetable for a class/section","operationId":"getByClassAndSection","parameters":[{"name":"studentClass","in":"path","required":true,"schema":{"type":"string"}},{"name":"section","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TimetableDto"}}}}}}}},"/api/v1/students/my-children":{"get":{"tags":["Students"],"summary":"List the authenticated parent's own children","operationId":"myChildren","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StudentDto"}}}}}}}},"/api/v1/homework/{studentClass}/{section}":{"get":{"tags":["Homework"],"summary":"List homework for a class/section (paginated)","operationId":"getByClassAndSection_1","parameters":[{"name":"studentClass","in":"path","required":true,"schema":{"type":"string"}},{"name":"section","in":"path","required":true,"schema":{"type":"string"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageHomeworkDto"}}}}}}},"/api/v1/fees/student/{studentId}":{"get":{"tags":["Fees"],"summary":"View fee status for a student","operationId":"getByStudent","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FeeDto"}}}}}}}},"/api/v1/exam-results/student/{studentId}":{"get":{"tags":["Exam Results"],"summary":"View exam results for a student","operationId":"getByStudent_1","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ExamResultDto"}}}}}}}},"/api/v1/auth/me":{"get":{"tags":["Auth"],"summary":"Get the currently authenticated user's profile","operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/v1/attendance/student/{studentId}":{"get":{"tags":["Attendance"],"summary":"View attendance history for a student","operationId":"getByStudent_2","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AttendanceDto"}}}}}}}},"/api/v1/attendance/class/{studentClass}/{section}/{date}":{"get":{"tags":["Attendance"],"summary":"View attendance for a whole class/section on a given date","operationId":"getByClassSectionDate","parameters":[{"name":"studentClass","in":"path","required":true,"schema":{"type":"string"}},{"name":"section","in":"path","required":true,"schema":{"type":"string"}},{"name":"date","in":"path","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AttendanceDto"}}}}}}}}},"components":{"schemas":{"UserCreateRequest":{"required":["email","name","password","role"],"type":"object","properties":{"name":{"maxLength":150,"minLength":0,"type":"string"},"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"},"role":{"type":"string","enum":["ADMIN","TEACHER","PARENT"]},"phone":{"maxLength":20,"minLength":0,"type":"string"}}},"UserDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["ADMIN","TEACHER","PARENT"]},"phone":{"type":"string"}}},"TimetableCreateRequest":{"required":["dayOfWeek","section","studentClass","subject","teacherId"],"type":"object","properties":{"studentClass":{"type":"string"},"section":{"type":"string"},"dayOfWeek":{"type":"string","enum":["MONDAY","TUESDAY","WEDNESDAY","THURSDAY","FRIDAY","SATURDAY","SUNDAY"]},"period":{"maximum":12,"minimum":1,"type":"integer","format":"int32"},"subject":{"type":"string"},"teacherId":{"type":"string","format":"uuid"}}},"TimetableDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"studentClass":{"type":"string"},"section":{"type":"string"},"dayOfWeek":{"type":"string","enum":["MONDAY","TUESDAY","WEDNESDAY","THURSDAY","FRIDAY","SATURDAY","SUNDAY"]},"period":{"type":"integer","format":"int32"},"subject":{"type":"string"},"teacherId":{"type":"string","format":"uuid"}}},"StudentCreateRequest":{"required":["dob","name","rollNo","section","studentClass"],"type":"object","properties":{"name":{"type":"string"},"rollNo":{"type":"string"},"studentClass":{"type":"string"},"section":{"type":"string"},"dob":{"type":"string","format":"date"},"parentId":{"type":"string","format":"uuid"}}},"StudentDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"rollNo":{"type":"string"},"studentClass":{"type":"string"},"section":{"type":"string"},"dob":{"type":"string","format":"date"},"parentId":{"type":"string","format":"uuid"}}},"NoticeCreateRequest":{"required":["targetRole","title"],"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"targetRole":{"type":"string","enum":["ADMIN","TEACHER","PARENT","ALL"]}}},"NoticeDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"title":{"type":"string"},"description":{"type":"string"},"targetRole":{"type":"string","enum":["ADMIN","TEACHER","PARENT","ALL"]},"createdBy":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"}}},"HomeworkCreateRequest":{"required":["dueDate","section","studentClass","subject","title"],"type":"object","properties":{"studentClass":{"type":"string"},"section":{"type":"string"},"subject":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"dueDate":{"type":"string","format":"date"}}},"HomeworkDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"studentClass":{"type":"string"},"section":{"type":"string"},"subject":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"dueDate":{"type":"string","format":"date"},"createdBy":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"}}},"ExamResultCreateRequest":{"required":["examName","marksObtained","maxMarks","studentId","subject","term"],"type":"object","properties":{"studentId":{"type":"string","format":"uuid"},"subject":{"type":"string"},"examName":{"type":"string"},"marksObtained":{"minimum":0,"exclusiveMinimum":false,"type":"number"},"maxMarks":{"minimum":0,"exclusiveMinimum":true,"type":"number"},"term":{"type":"string"}}},"ExamResultDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"studentId":{"type":"string","format":"uuid"},"subject":{"type":"string"},"examName":{"type":"string"},"marksObtained":{"type":"number"},"maxMarks":{"type":"number"},"grade":{"type":"string"},"term":{"type":"string"}}},"RefreshRequest":{"required":["refreshToken"],"type":"object","properties":{"refreshToken":{"type":"string"}}},"AuthResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"role":{"type":"string","enum":["ADMIN","TEACHER","PARENT"]}}},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"AttendanceMarkRequest":{"required":["date","status","studentId"],"type":"object","properties":{"studentId":{"type":"string","format":"uuid"},"date":{"type":"string","format":"date"},"status":{"type":"string","enum":["PRESENT","ABSENT","LATE","EXCUSED"]}}},"AttendanceDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"studentId":{"type":"string","format":"uuid"},"date":{"type":"string","format":"date"},"status":{"type":"string","enum":["PRESENT","ABSENT","LATE","EXCUSED"]},"markedBy":{"type":"string","format":"uuid"}}},"StudentUpdateRequest":{"type":"object","properties":{"name":{"type":"string"},"rollNo":{"type":"string"},"studentClass":{"type":"string"},"section":{"type":"string"},"dob":{"type":"string","format":"date"},"parentId":{"type":"string","format":"uuid"}}},"FeeUpdateRequest":{"type":"object","properties":{"amountPaid":{"minimum":0,"exclusiveMinimum":false,"type":"number"},"status":{"type":"string","enum":["PENDING","PARTIAL","PAID","OVERDUE"]}}},"FeeDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"studentId":{"type":"string","format":"uuid"},"term":{"type":"string"},"amountDue":{"type":"number"},"amountPaid":{"type":"number"},"status":{"type":"string","enum":["PENDING","PARTIAL","PAID","OVERDUE"]},"dueDate":{"type":"string","format":"date"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"PageUserDto":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"unpaged":{"type":"boolean"},"offset":{"type":"integer","format":"int64"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"pageSize":{"type":"integer","format":"int32"},"pageNumber":{"type":"integer","format":"int32"},"paged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"direction":{"type":"string"},"nullHandling":{"type":"string"},"ascending":{"type":"boolean"},"property":{"type":"string"},"ignoreCase":{"type":"boolean"}}},"PageStudentDto":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/StudentDto"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageNoticeDto":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/NoticeDto"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageHomeworkDto":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/HomeworkDto"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}