# src/main/java/com/school/app/student/StudentController.java
package com.school.app.student;

import com.school.app.user.User;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.UUID;

@RestController
@RequestMapping("/api/v1/students")
@RequiredArgsConstructor
@Tag(name = "Students")
public class StudentController {

    private final StudentService studentService;

    @GetMapping
    @PreAuthorize("hasAnyRole('ADMIN', 'TEACHER')")
    @Operation(summary = "List students (paginated)")
    public Page<StudentDto> list(Pageable pageable) {
        return studentService.list(pageable);
    }

    @GetMapping("/my-children")
    @PreAuthorize("hasRole('PARENT')")
    @Operation(summary = "List the authenticated parent's own children")
    public List<StudentDto> myChildren(@AuthenticationPrincipal User currentUser) {
        return studentService.getMyChildren(currentUser);
    }

    @PostMapping
    @PreAuthorize("hasRole('ADMIN')")
    @Operation(summary = "Create a student")
    public StudentDto create(@Valid @RequestBody StudentCreateRequest request) {
        return studentService.create(request);
    }

    @GetMapping("/{id}")
    @PreAuthorize("hasAnyRole('ADMIN', 'TEACHER', 'PARENT')")
    @Operation(summary = "Get a student by id (parents may only view their own child)")
    public StudentDto getById(@PathVariable UUID id, @AuthenticationPrincipal User currentUser) {
        return studentService.getById(id, currentUser);
    }

    @PatchMapping("/{id}")
    @PreAuthorize("hasRole('ADMIN')")
    @Operation(summary = "Update a student")
    public StudentDto update(@PathVariable UUID id, @Valid @RequestBody StudentUpdateRequest request) {
        return studentService.update(id, request);
    }
}